Finalize everything
