+++ packages/mobile/package.json
@@ -74,6 +74,7 @@
     "bn.js": "5.2.0",
     "bs58": "5.0.0",
     "color": "3.2.1",
+    "deprecated-react-native-prop-types": "4.0.0",
     "expo-crypto": "9.2.0",
     "ffmpeg-kit-react-native": "4.5.2",
     "formik": "2.2.9",
@@ -102,6 +103,7 @@
     "react-native-document-picker": "^8.1.2",
     "react-native-draggable-flatlist": "3.0.7",
     "react-native-dynamic": "1.0.0",
+    "react-native-fast-crypto": "2.2.0",
     "react-native-fs": "2.18.0",
     "react-native-gesture-handler": "1.10.3",
     "react-native-get-random-values": "1.8.0",
