+++ packages/mobile/src/services/audius-backend-instance.ts
@@ -3,11 +3,11 @@ import * as nativeLibs from '@audius/sdk/dist/native-libs'
 import AsyncStorage from '@react-native-async-storage/async-storage'
 import { Image } from 'react-native'
 import Config from 'react-native-config'
-// import scrypt from 'react-native-scrypt'
 
 import { track } from 'app/services/analytics'
 import { reportToSentry } from 'app/utils/reportToSentry'
 
+import { createPrivateKey } from './createPrivateKey'
 import { withEagerOption } from './eagerLoadUtils'
 import { env } from './env'
 import {
@@ -19,42 +19,6 @@ import {
 import { monitoringCallbacks } from './monitoringCallbacks'
 import { getFeatureEnabled } from './remote-config'
 import { remoteConfigInstance } from './remote-config/remote-config-instance'
-
-function bufferFromHexString(hexString: string) {
-  const byteArray = hexString
-    .match(/.{1,2}/g)
-    ?.map((byte) => parseInt(byte, 16))
-  return new Uint8Array(byteArray as number[])
-}
-
-/**
- * Given a user encryptStr and initialization vector, generate a private key
- * @param encryptStr String to encrypt (can be user password or some kind of lookup key)
- * @param ivHex hex string iv value
- */
-const createKey = async (encryptStr: string, ivHex: string) => {
-  const N = 32768
-  const r = 8
-  const p = 1
-  const dkLen = 32
-  const encryptStrBuffer = Buffer.from(encryptStr)
-  const ivBuffer = Buffer.from(ivHex)
-
-  // const derivedKey = await scrypt(
-  //   encryptStrBuffer,
-  //   ivBuffer,
-  //   N,
-  //   r,
-  //   p,
-  //   dkLen,
-  //   'buffer'
-  // )
-  const keyHex = ''
-
-  // This is the private key
-  const keyBuffer = bufferFromHexString(keyHex)
-  return { keyHex, keyBuffer }
-}
 
 /**
  * audiusBackend initialized for a mobile environment
@@ -85,7 +49,7 @@ export const audiusBackendInstance = audiusBackend({
     )
   }),
   hedgehogConfig: {
-    createKey
+    createKey: createPrivateKey
   },
   identityServiceUrl: Config.IDENTITY_SERVICE,
   generalAdmissionUrl: Config.GENERAL_ADMISSION,
